support comma-separated paths in Go-Import-Path
Some packages can be imported from several import paths.
XS-Go-Import-Path can reflect this by using a comma-separated list
of known import paths for the package.

This patch updates make.go's existing package detection mechanism to
support this.

The pkgs struct field ImportPath is renamed to XSGoImportPath to avoid
confusion because it could be that XSGoImportPath is not a valid import
path if it contains commas.

For example, the following format would now be valid:
```
XS-Go-Import-Path: github.com/go-mgo/mgo,
                   gopkg.in/mgo.v2,
                   labix.org/v2/mgo,
                   launchpad.net/mgo
```