Download only the repo, not its dependencies.
Thanks to tincho for the hint!

This refactors standard library filtering to use “go list std” instead
of using “go list” to tell us about which packages are in the standard
library (which would require the packages to be present).

This also means we don’t need to run “go get” after removing vendor/
anymore.

Further, we now use vcs.CreateAtRev instead of running “git reset”
ourselves, which is slightly cleaner and VCS-agnostic (all but 5
upstreams packaged in Debian do use git, though…).

This commit cuts down cloning time of large packages to
1/3rd (e.g. grafana goes from 3 minutes to 1 minute).