switch to github.com/google/go-github/github
With this change, the GITHUB_USERNAME, GITHUB_PASSWORD and GITHUB_OTP
environment variables are introduced. Authenticating to GitHub is useful to
obtain a higher rate limit.

Furthermore, an in-memory cache is now used to reduce the number of requests
sent to GitHub in the first place.

fixes https://bugs.debian.org/898397