Make sure to call normalizeDebianProgramName() on the whole name, not on parts
According to policy, the minimum length for a package is two, which is
something normalizeDebianProgramName() checks. However this function was
called on *parts* of the package name, and in case of a package such as
golang-golang-x-term, it fails, as the part "x" is only one character.

Note that we can safely drop strings.Trim(), as the string returned by
normalizeDebianProgramName() was already trimmed from '-'

Closes: #148