Fix make output: run 'git push' before 'gbp push'
As mentiond in #107, running 'gbp push' when there's no debian tag is
not the right thing to do,  as it will only push the upstream branch.
Consequences are:

1. The default gitlab branch is set to the upstream branch instead of
   the debian branch.
2. The debian branch is not pushed and needs to be pushed manually
   anyway.

I believe that the best thing to do instead is just to run 'git push
origin <debian-branch>' first, and then run 'gbp push'.

Closes: #107