Extend default debian/gbp.conf with extra security config tips
When creating a new package, populate the git-buildpackage with additional
configs and in-line comments on why and how to use them. This will make
go packaging easier, more consistent and more secure as the best practices
flow to all packages via good defaults.

Contents is in line with the template used by `dh-make` version 2.202503.