dh-make-golang make: directly use the user preferred tag specified in "-git_revision" option
When a user specifies a valid tag when using

    dh-make-golang make -git_revision {some valid tag}

if there is another tag pointing to the same commit as the specified
tag, then dh-make-golang might use the other tag to determine the
package version.

This patch modifies the behavior of "dh-make-golang make" such that, if
the user specifies a valid tag in "-git_revision" option,
"dh-make-golang make" will always use the specified tag to determine the
upstream package version.

This patch also trivially changes the tests written in version_test.go
so that the tests adjusts to the new function signature of
version.go:pkgVersionFromGit().

Signed-off-by: William Lyu <William.Lyu@windriver.com>