fix release tags not recognized when not on a branch
"dh-make-golang" creates the "debianized" source repo from scratch. It
creates an empty git repo, adds the upstream as a remote, and fetches
the tags. However, before this patch, it is using "git fetch" with no
additional flags or options. This results in upstream tags not on any
branches not being picked up. This patch adds "--tags" flag to
"git fetch" so that it can fetch all upstream tags.

Signed-off-by: William Lyu <William.Lyu@windriver.com>