Fix dh-make-golang estimate
As support for the gopath mode of "go get" has been removed in latest
versions of Go, we cannot use the GO111MODULE=off trick any more.
To fix the estimate command, we now download the sources of the repo
manually then run "go get" in the repo dir to make use of the
information contained in the go.mod file and download the dependencies.

For go packages that are not modules yet, we init ourselves a go module
using "go mod init".