Make sure to remove all temporary directories
Previously, some files could not be removed because of the directories
permissions, which made os.RemoveAll() fail silently and left a lot of
garbage files in /tmp.

This commit adds a new forceRemoveAll() function that removes all files
and directories more reliably, in a new fs_utils.go file for filesystem
utilities. isFile() is also moved to this new file.

Also log errors when forceRemoveAll fails to detect these issues sooner
in the future.