Check others major versions of packages in estimate
Instead of immediately checking that the repo root path is in the
list of packages in Debian, use the full import path (including the
potential major version suffix), then check for other major versions
if applicable, and finally, check for the repo root.

This allows to display that another major version is currently packaged
in Debian, and prevents hiding necessary major upgrades to already
packaged Go modules with a mismatched version.