Show and indicate duplicate dependencies in estimate
Previously, when a dependency was seen multiple times, it was
deduplicated. This helped seeing the exact number of modules that needed
to be packaged, but it didn't help seeing the most modules to package in
priority.

Now, we show duplicate packages but in a dimmed color and with a count
of occurences in parenthesis at the end. The colors allows to still
quickly evaluate the effort by focusing the eyes on the deduplicated
list, while the count helps detecting the most used module.