Only highlight unique repositories in estimate
Keep printing all needed modules recursively, but instead of only
dimming the duplicated modules, we also dim the rest of the import path
after the repo root, and modules whose repo root has already been
printed. This allows to only highlight the repository names (the actual
thing that we package in Debian) and only once, so it give a better
appreciation of the work packaging work required.