Revert "Put build artifacts in `debian/build/` to avoid polluting upstream sources"
This reverts commit 0bdd05c4917581df680d7489a3659e29fec46676.

While the general idea makes sense and would make things better, this
had some push back on the review and still got merged anyway. It also
conflicts with an old proposal that still is pending to be updated
and re-proposed to try to unify the build paths for all packages at the
dpkg and debhelper layer, via a new dpkg-build-api level. And doing a
migration twice seems wasteful.

Ref: https://lists.debian.org/debian-dpkg/2020/03/msg00003.html