estimate: do not print children of mismatched major mods
Simply log a message once about mismatched major versions in the log
output, and do not print all of their children.

For packages with a lot of dependencies, the previous behaviour was
producing a lot of noise. If a module is already packaged, even with
a different major version, it will probably not miss a lot of
dependencies.