estimate: fix occasional infinite recursion
The "go mod graph" command returns a DAG so there can be no cycles, but
modules can appear multiple times with different versions. When
stripping the version from the modules we can accidentaly introduce
cycles in the DAG by merging multiple versions of a module.

To prevent this, we go back to saving the exact output of "go mod graph"
in memory, including the version strings, and get rid of them when
visiting each node.

Closes: #292